requirements.txt
Django>=4.1
djangorestframework
djangorestframework-simplejwt
python-telegram-bot==13.7
psycopg2-binary
requests